+++ js/ui/interaction.js
@@ -138,8 +138,11 @@ function Interaction(el) {
         }
     }
 
-    function doubleclick(point) {
-        interaction.fire('dblclick', {point: point});
+    function doubleclick(point, ev) {
+        interaction.fire('dblclick', {
+            point: point,
+            originalEvent: ev
+        });
     }
 
     function onmousedown(ev) {
@@ -206,7 +209,7 @@ function Interaction(el) {
     }
 
     function ondoubleclick(ev) {
-        doubleclick(mousePos(ev));
+        doubleclick(mousePos(ev), ev);
         zoom('wheel', Infinity * (ev.shiftKey ? -1 : 1), mousePos(ev));
         ev.preventDefault();
     }